# lib/api-zod/src/generated/types/createTeamMemberRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTeamMemberRequest {
  category: string;
  name: string;
  role: string;
  location?: string;
  bio?: string;
  email?: string;
  linkedin?: string;
  websiteUrl?: string;
  sortOrder?: number;
}